chore: refactor vaultsList
