+++ docs/upgrades/data-and-component-cleanup/step-2-1-strategies-decomposition.md
@@ -13,9 +13,11 @@ Successfully decomposed the strategies-panel component into smaller, focused com
 ### 1. Component Breakdown
 
 **Original File:**
+
 - `strategies-panel.tsx` - 560 lines (after Step 1.1 data extraction)
 
 **New Component Structure:**
+
 ```
 ### 🔧 **New Component Structure**
 
@@ -29,11 +31,13 @@ Created a clean component hierarchy in `/src/components/strategies-panel/`:
 ```
 
 **Wrapper Component:**
+
 - Wrapper component removed - direct imports now used
 
 ### 2. Component Responsibilities
 
 #### **StrategiesPanel** (Main Container)
+
 - **Purpose:** Orchestrates child components and manages tab navigation
 - **Responsibilities:**
   - Tab state management
@@ -43,6 +47,7 @@ Created a clean component hierarchy in `/src/components/strategies-panel/`:
 - **Props:** `vaultChainId`, `vaultAddress`, `vaultDetails`
 
 #### **StrategyAllocationChart** (Chart Display)
+
 - **Purpose:** Renders allocation and APY contribution pie charts
 - **Responsibilities:**
   - Pie chart rendering with Recharts
@@ -52,6 +57,7 @@ Created a clean component hierarchy in `/src/components/strategies-panel/`:
 - **Props:** `allocationData`, `apyContributionData`, `totalAPYContribution`, `colors`
 
 #### **StrategyTable** (Table Container)
+
 - **Purpose:** Manages table structure and sorting controls
 - **Responsibilities:**
   - Table header with sortable columns
@@ -61,6 +67,7 @@ Created a clean component hierarchy in `/src/components/strategies-panel/`:
 - **Props:** Strategy arrays, sorting state, event handlers
 
 #### **StrategyRow** (Individual Row)
+
 - **Purpose:** Renders individual strategy data and details
 - **Responsibilities:**
   - Strategy information display
@@ -72,16 +79,19 @@ Created a clean component hierarchy in `/src/components/strategies-panel/`:
 ### 3. Architectural Improvements
 
 #### **Separation of Concerns**
+
 - ✅ **UI Logic:** Cleanly separated into focused components
 - ✅ **Business Logic:** Remains in custom hooks (useStrategiesData, useSortingAndFiltering)
 - ✅ **Presentation:** Each component has a single, clear responsibility
 
 #### **Reusability**
+
 - ✅ **StrategyRow:** Can be reused in other strategy-related interfaces
 - ✅ **StrategyAllocationChart:** Reusable for any allocation visualization
 - ✅ **StrategyTable:** Can handle different strategy datasets
 
 #### **Maintainability**
+
 - ✅ **Smaller Files:** Easier to navigate and understand
 - ✅ **Clear Interfaces:** Well-defined props make components predictable
 - ✅ **Type Safety:** Full TypeScript coverage maintained
@@ -89,13 +99,16 @@ Created a clean component hierarchy in `/src/components/strategies-panel/`:
 ### 4. Performance Optimizations
 
 #### **React.memo() Implementation**
+
 Applied `React.memo()` to all major components to prevent unnecessary re-renders:
+
 - `StrategiesPanel` - Memoized with props comparison
 - `StrategyTable` - Memoized to prevent table re-renders
 - `StrategyRow` - Memoized for individual row optimization
 - `StrategyAllocationChart` - Memoized for chart re-render prevention
 
 #### **Component Isolation**
+
 - Each component only re-renders when its specific props change
 - Chart components isolated from table state changes
 - Row expansion state doesn't affect other rows
@@ -103,6 +116,7 @@ Applied `React.memo()` to all major components to prevent unnecessary re-renders
 ## Files Modified
 
 ### New Components Created
+
 1. `/src/components/strategies-panel/StrategiesPanel.tsx` - 132 lines
 2. `/src/components/strategies-panel/StrategyAllocationChart.tsx` - 159 lines  
 3. `/src/components/strategies-panel/StrategyTable.tsx` - 111 lines
@@ -110,23 +124,27 @@ Applied `React.memo()` to all major components to prevent unnecessary re-renders
 5. `/src/components/strategies-panel/index.ts` - 4 lines
 
 ### Files Refactored
+
 1. `/src/routes/vaults/$chainId/$vaultAddress/index.tsx` - Updated imports to use direct component
 2. **Removed:** `/src/components/strategies-panel.tsx` - Wrapper component eliminated
 
 ## Technical Implementation Details
 
 ### Component Props Design
+
 - **Minimal Props:** Each component receives only what it needs
 - **Event Handlers:** Clean callback pattern for user interactions
 - **Data Props:** Properly typed data structures from custom hooks
 - **Configuration Props:** Optional props for customization (colors, flags)
 
 ### TypeScript Integration
+
 - **Interface Definitions:** Clear interfaces for all component props
 - **Type Safety:** Full type coverage with proper imports from data types
 - **Export Strategy:** Barrel exports for clean import statements
 
 ### Chart Component Features
+
 - **Custom Tooltips:** Rich tooltip display with formatted data
 - **Responsive Design:** Adapts to mobile and desktop layouts
 - **Color Theming:** Configurable color schemes
@@ -135,17 +153,20 @@ Applied `React.memo()` to all major components to prevent unnecessary re-renders
 ## Impact Assessment
 
 ### Code Organization Improvements
+
 - ✅ **Single Responsibility:** Each component has one clear purpose
 - ✅ **Easier Testing:** Components can be tested in isolation
 - ✅ **Better Debugging:** Issues can be traced to specific components
 - ✅ **Team Development:** Multiple developers can work on different components
 
 ### Bundle Size Impact
+
 - ✅ **Code Splitting Ready:** Components can be lazy-loaded if needed
 - ✅ **Tree Shaking:** Unused components can be eliminated
 - ✅ **No Size Increase:** Decomposition didn't add overhead
 
 ### Performance Impact
+
 - ✅ **Render Optimization:** React.memo prevents unnecessary re-renders
 - ✅ **Isolation Benefits:** Changes to one component don't affect others
 - ✅ **Memory Usage:** Better garbage collection of unused component instances
@@ -153,12 +174,14 @@ Applied `React.memo()` to all major components to prevent unnecessary re-renders
 ## Validation
 
 ### Build Tests
+
 - ✅ TypeScript compilation successful
 - ✅ ESLint rules passing
 - ✅ No runtime errors
 - ✅ Bundle size within acceptable limits
 
 ### Functional Tests
+
 - ✅ All strategy table functionality preserved
 - ✅ Chart rendering working correctly
 - ✅ Row expansion/collapse functional
@@ -166,6 +189,7 @@ Applied `React.memo()` to all major components to prevent unnecessary re-renders
 - ✅ External links working properly
 
 ### Component Interface Tests
+
 - ✅ Props properly typed and validated
 - ✅ Event handlers firing correctly
 - ✅ Data flow working as expected
