+++ src/components/vaults-list/index.ts
@@ -1,0 +1,4 @@
+export { VaultsList } from './VaultsList'
+export { VaultsTable } from './VaultsTable'
+export { VaultsTableHeader } from './VaultsTableHeader'
+export { VaultsSearchBar } from './VaultsSearchBar'
