+++ src/routes/index.tsx
@@ -1,5 +1,5 @@
 import { createFileRoute } from '@tanstack/react-router'
-import VaultsList from '../components/VaultsList'
+import { VaultsList } from '@/components/vaults-list'
 import { useVaults } from '@/contexts/useVaults'
 import { Vault } from '@/types/vaultTypes'
 import YearnLoader from '@/components/YearnLoader'
